# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BTI Smart Attendance Monitoring System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects, optionally filtered by branch
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: section
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    Subject:
      type: object
      required: [id, name, code, branchId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        createdAt:
          type: string
    SubjectInput:
      type: object
      required: [name, code, branchId]
      properties:
        name:
          type: string
        code:
          type: string
        branchId:
          type: integer
    SubjectUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        branchId:
          type: integer
    Student:
      type: object
      required: [id, studentId, name, branchId, year, section, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: string
        name:
          type: string
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        year:
          type: integer
        section:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [studentId, name, branchId, year, section]
      properties:
        studentId:
          type: string
        name:
          type: string
        branchId:
          type: integer
        year:
          type: integer
        section:
          type: string
        phone:
          type: string
        email:
          type: string
    StudentUpdate:
      type: object
      properties:
        studentId:
          type: string
        name:
          type: string
        branchId:
          type: integer
        year:
          type: integer
        section:
          type: string
        phone:
          type: string
        email:
          type: string
...