# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BTI Smart Attendance Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  studentId: string;
  name: string;
  branchId: number;
  /** @nullable */
  branchName?: string | null;
  year: number;
  section: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}