# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BTI Smart Attendance Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  studentId: string;
  name: string;
  branchId: number;
  year: number;
  section: string;
  phone?: string;
  email?: string;
}