# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BTI Smart Attendance Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  name?: string;
  code?: string;
  branchId?: number;
}